Rename to PackageJsonReader
