+++ apps/cli/src/adapters/commander/commands/doctor.ts
@@ -2,11 +2,11 @@ import { Command } from "commander";
 import * as process from "node:process";
 
 import { Dependencies } from "../../../domain/dependencies.js";
-import { checkMonorepoScripts } from "../../../domain/node.js";
+import { checkMonorepoScripts } from "../../../domain/package-json.js";
 
 type DoctorDependencies = Pick<
   Dependencies,
-  "logger" | "nodeReader" | "repositoryReader"
+  "logger" | "packageJsonReader" | "repositoryReader"
 >;
 
 export const makeDoctorCommand = (dependencies: DoctorDependencies): Command =>
