+++ apps/cli/src/domain/__tests__/data.ts
@@ -1,11 +1,11 @@
 import { mock } from "vitest-mock-extended";
 
 import { Logger } from "../logger.js";
-import { NodeReader } from "../node.js";
+import { PackageJsonReader } from "../package-json.js";
 import { RepositoryReader } from "../repository.js";
 
 export const makeMockDependencies = () => ({
   logger: mock<Logger>(),
-  nodeReader: mock<NodeReader>(),
+  packageJsonReader: mock<PackageJsonReader>(),
   repositoryReader: mock<RepositoryReader>(),
 });
