+++ apps/cli/src/index.ts
@@ -1,12 +1,12 @@
 import { makeCli } from "./adapters/commander/index.js";
 import { makeConsoleLogger } from "./adapters/console/index.js";
-import { makeNodeReader } from "./adapters/node/index.js";
+import { makePackageJsonReader } from "./adapters/node/package-json.js";
 import { makeRepositoryReader } from "./adapters/node/repository.js";
 import { Dependencies } from "./domain/dependencies.js";
 
 const deps: Dependencies = {
   logger: makeConsoleLogger(),
-  nodeReader: makeNodeReader(),
+  packageJsonReader: makePackageJsonReader(),
   repositoryReader: makeRepositoryReader(),
 };
